refactor(view) : 뷰 메시지 수정
- 입력 형식 및 예시 추가
